Rework data fetching and add new sensors
This is quite a substantial update that completely changes how data are
loaded from the datasource. It uses official SOAP API instead of relying
on internals of web UI. It no longer needs to fetch CNB rate manually as
the API can convert it for us.

The data are now fetched in coordinator instead of in every sensor
update, which should reduce resource usage.

New sensors are added for easier automations.